WIP: Refactor FileManager
