+++ dkist/io/asdf/tests/test_dataset.py
@@ -12,8 +12,8 @@ from asdf.testing.helpers import roundtrip_object
 
 import dkist
 from dkist.data.test import rootdir
-from dkist.io import FileManager
-from dkist.io.loaders import AstropyFITSLoader
+from dkist.io import DKISTFileManager
+from dkist.io.dask.loaders import AstropyFITSLoader
 
 
 @pytest.fixture
@@ -26,8 +26,8 @@ def tagobj(request):
 
 @pytest.fixture
 def file_manager():
-    return FileManager.from_parts(["test1.fits", "test2.fits"], 0, "float", (10, 10),
-                                  loader=AstropyFITSLoader)
+    return DKISTFileManager.from_parts(["test1.fits", "test2.fits"], 0, "float", (10, 10),
+                                       loader=AstropyFITSLoader)
 
 
 def test_roundtrip_file_manager(file_manager):
